Trivially relocatable (1/3): Traits and algorithms.
This implements the library behavior proposed in D1144R6.
This patch doesn't do any special optimizations (e.g. to `std::vector`);
those will come later. This is just the bare minimum to conform to
P1144's additions to the standard.

We mangle all the P1144 names behind `__libcpp_...`, so that they
aren't taken by the user-programmer to be official Standard C++ names.

    struct is_relocatable<T>
    struct is_nothrow_relocatable<T>
    struct is_trivially_relocatable<T>

    bool is_relocatable_v<T>
    bool is_nothrow_relocatable_v<T>
    bool is_trivially_relocatable_v<T>

    T relocate(T*)
    T* relocate_at(T*, T*)
    InputIt uninitialized_relocate(InputIt, InputIt, ForwardIt)
    pair<InputIt, ForwardIt> uninitialized_relocate_n(InputIt, Size, ForwardIt)