[libc++] Use `__libcpp_memswap` in `swap_ranges` when possible
This follows the same pattern used for `copy`, `copy_backward`, and `move`,
except that `swap_ranges` requires four parameters instead of three,
and we need to introduce the `__can_consteval` flag so that `swap_ranges`
can set it to `false` in its trivial codepath.

The new tests mainly verify that the new codepaths are correctly disabled
during constant evaluation. When I wrote them, I was worried that the
trivial codepath might try to form invalid pointer values. But in fact
I believe that the trivial codepath is actually never entered unless
the sentinel is an iterator that unwraps to a pointer type; so we should
be okay in that respect.