[libc++] Introduce __libcpp_memswap; use it for swapping built-in arrays
Swapping built-in arrays of trivially copyable types can be done
using bytewise swapping ("memswap", by analogy to memcpy and memmove).
We choose a 128-byte stack buffer because that's the largest power-of-2
where Clang trunk will inline the __builtin_memcpy; this is not a
scientific benchmark-based approach at all. Fortunately the buffer size
is easy to change, if anyone wants to change it later.

`__libcpp_memswap` depends on casting from `void*` back to `char*`
in order to do the pointer math. Therefore it is never constexpr.
(This will become important, later, when we start trying to use
`__libcpp_memswap` in STL algorithms that might be constant-evaluated.)