[llvm] IFSStub destructor should be virtual
The problem in this code is that IFSStub is being used as a polymorphic
base class according to the basic pattern

    IFSStub *p = new IFSStubTriple;
    delete p; // UB, fails to destroy most-derived object

The problem is disguised by the use of `unique_ptr`. It was
detected by patching the standard library to prevent implicit conversion
between `unique_ptr<Derived>` and `unique_ptr<Base>` unless `Base` has
a virtual destructor.

The implicit conversion from `unique_ptr<IFSStubTriple>` to `unique_ptr<IFSStub>`
is happening on this line, in `ifs::readIFSFromBuffer`:

    llvm/lib/InterfaceStub/IFSHandler.cpp:204:10:
      return std::move(Stub);
             ^~~~~~~~~~~~~~~

`unique_ptr<IFSStubTriple> Stub`'s destructor will correctly call
the destructor of `IFSStubTriple`; but after the implicit conversion,
we have only a `unique_ptr<IFSStub>`, which will incorrectly call only
the destructor of the base class `IFSStub`. The incorrect base-class
destructor happens on this line:

    llvm/unittests/InterfaceStub/ELFYAMLTest.cpp:46
      Expected<std::unique_ptr<IFSStub>> StubOrErr = readIFSFromBuffer(Data);
      [...]
      std::unique_ptr<IFSStub> Stub = std::move(StubOrErr.get());
      [...]
    } // destroy base class here, UB

The simplest fix is to make that base-class destructor virtual.