[clang] __is_trivially_equality_comparable for types containing lambdas
Lambdas (closure types) are trivially equality-comparable iff they are
non-capturing, because non-capturing lambdas are convertible to function
pointers: if `(lam1 == lam2)` compiles, then `lam1` and `lam2` must have
the same type, and be always-equal, and be empty.
...Or else `lam1` and/or `lam2` must inherit from a type like that, but
not necessarily be exactly that type. That is, lambdas still have the
same problems as any other class type w.r.t. base classes (#63192).
All this commit does is teach the compiler about the special case where
we know how lambda comparison works (teaching that it's not a problem).