[libc++] Avoid using std::forward on things that aren't forwarding references
See https://quuxplusone.github.io/blog/2023/05/27/dont-forward-non-forwarding-references/
After this patch, it's possible to mechanically search-and-replace like this:
`s/[_A-Za-z]*::forward<.*>[(](.*)[)]/_LIBCPP_FWD(\1)/g`
and then `#define _LIBCPP_FWD(x) decltype(x)(x)`.
We don't want to actually land that search-and-replace on trunk, but it's a nice
property to have: to use `forward` only when we're actually forwarding, and not
in any other situation.

Differential Revision: https://reviews.llvm.org/D120998