[libc++] [P3016R1] Remove free begin/end/data/empty from <valarray> and <initializer_list>
Put the change in only for C++26. It's actually tricky to put in globally,
because in C++14 mode the old std::begin(initializer_list<E>) was constexpr-friendly
but the generic std::begin(C&) wasn't constexpr-friendly; therefore the latter
can't substitute for the former in C++14. We could probably do it in C++17
and later; but as long as we're #ifdef'ing it at all, let's do it in the
style that libc++ trunk would eventually end up doing it.

I believe it would be difficult (i.e. require compiler magic that we don't
currently have) to give a deprecation warning for the declarations in
`<initializer_list>` but _not_ the declarations in `<iterator>`.

Note -- this is just the little leftover bits that weren't already implemented
by Jiang An in trunk. It'll go away completely soon, I expect.