[libc++] [P3016R0] Remove free begin/end from <valarray> and <initializer_list>
Casey Carter reports that this program fails to compile because std::cbegin
wants to call std::begin, and the std::begin overload that takes `valarray`
isn't declared yet at the time we process std::cbegin's primary template.
(There is a std::begin overload that takes "anything with begin/end members,"
but `valarray` doesn't have begin/end members... before now, that is.)

    #include <iterator>
    #include <valarray>
    int main() {
      std::valarray<int> v{1,2,3};
      (void)std::cbegin(v);
      (void)std::cend(v);
    }

Does removing the overloads from <initializer_list> break these usages?
    std::begin({1,2,3});
    for (int i : {1,2,3}) {}
We don't want to break the latter!
The former we do break, but maybe that's a feature not a bug, because
people HAVE gotten confused about why std::begin works and std::cbegin
doesn't: https://stackoverflow.com/questions/50395787/cant-define-an-initializer-list-in-cbegin
So making them both consistently not-work could be a plus.