[libc++] Simplify std::sort helpers
These helpers were introduced in https://reviews.llvm.org/D118029
and then `sort` itself was rewritten in https://reviews.llvm.org/D122780
and the helpers simplified in https://reviews.llvm.org/D144478 .

This patch further simplifies the helpers. We apply D144478's
optimization (to not count the number of swaps) to `__sort3` everywhere
it's used except for the one place in `nth_element` that requires the
return value.

We permit the branchless version to be used for pointers, e.g.
    int *a[10]; std::sort(a, a+10, std::less<int*>());
and we prohibit the branchless version from being used for unintentionally
expensive comparators, e.g.
    double a[10]; std::sort(a, a+10, std::less<BigDouble>());
    Derived *a[10]; std::sort(a, a+10, std::less<VirtualBase*>());
In both cases it would be safe, but more expensive than the original PR
intended (since the comparator will be invoked more often in the branchless
version).