[libc++] Implement _LIBCPP_NONCONFORMING_OPTIONAL_MOVE_CTOR
The Standard requires that `optional`'s move constructor
leave the right-hand side in a "moved-from, but still engaged" state,
even when it would be cheaper to leave it in a "disengaged" state.
We implement the latter behavior under a macro that can be turned
on and off. It's off by default.

This makes for inconsistent behavior, which is probably why C++17 chose
to make the move-constructor never disengage the right-hand side. See the
new test files for some examples.

This was proposed in [P3055R1], and then removed again in [P3055R2]
both to narrow the scope and because it's probably a bad idea to
change it this long after the ship has sailed.