[libc++] Implement _LIBCPP_NONCONFORMING_OPTIONAL_MOVE_CTOR
This is related to [P3055]. It's currently required that `optional`'s
move constructor leave the right-hand side in a "moved-from, but still
engaged" state, even when it would be cheaper to leave it in a
"disengaged" state. We implement the latter behavior under a macro
that can be turned on and off (it's off by default).

This makes for inconsistent behavior, which is probably why C++17 chose
to make the move-constructor never disengage the right-hand side. See the
new test files for some examples.