[clang] Implement the [[clang::trivially_relocatable]] attribute
This works like [[clang::trivial_abi]], in that it marks the class as
trivially relocatable regardless of its non-trivial special members.
That is, the attribute warrants to the compiler that the programmer
knows what they're doing (as in unique_ptr and vector).
However, [[clang::trivially_relocatable]] doesn't affect ABI such as
parameter-passing convention; it merely makes the builtin type trait
`__is_trivially_relocatable(T)` return true.