[clang] Rough pass at a [[clang::nrvo]] attribute
The idea of this attribute is that you put it on a return statement
to enforce NRVO.

    [[nrvo]] return x;   // OK, warns if x isn't RVOed
    [[nrvo]] return;     // Ill-formed
    [[nrvo]] return x+1; // Ill-formed

Ideally, the attribute would also be a hint to the optimizer that
the named variable ought to be preferred as the NRVO candidate for
this function/scope.

Ideally, the attribute would let you return non-movable types
via explicit NRVO. But that's probably really hard to teach Clang.