[clang] Support [[clang::foo(bar)]] in C++03 mode
The `__has_cpp_attribute` builtin previously gave the wrong answer
(0 instead of 1), which caused the parser to treat the attribute as
unrecognized instead of recognized; this meant that e.g.
`[[gnu::visibility("hidden")]]` would fail to parse in C++03 mode
because the parser wouldn't recognize `visibility` as an attribute
taking any arguments (because it would assume it was unrecognized).

Fix the tablegen and then update some attribute-related tests to run
in C++03 mode. Those tests didn't pass before this patch.