[libc++] Optimize vector insert/erase to use trivial relocation for pmr::vector<pmr::string>
But not insert/erase of vector<pmr::string>, because the allocators could differ.
And not insert/erase of vector<AggContainingPmrString>, because all bets
are off in that case.

This changes the observable behavior of my existing test which explicitly
tests the undefined-behavior cases (i.e. cases that break the PMR invariant
of a `pmr::vector`). No defined-behavior case is affected.