[clang] -Wassign-to-class-rvalue
Implement three new warnings:

(1) `-Wassign-to-class-rvalue`, which is part of `-Wmost` and therefore
part of `-Wall`. This covers assignment, compound assignment (e.g. `+=`),
and postincrement and postdecrement. But it is emitted at all only in C++23
and later, because in C++20 it gives false positives on library types
such as `tuple<int&>`. These types are const-assignable only in C++23
and later.

(2) `-Wpreincrement-of-class-rvalue`, which is off by default.
This covers preincrement and predecrement. These do happen a lot in
real life; specifically with iterator types. For example ASIO contains
expressions of both forms: `--end()` and `++begin()`. These could be
rewritten to `std::prev(end())` and `std::next(begin())`, but in practice
I doubt people really want to do that.

On the other hand, `begin()++` or `end()--` is back in the category
of always being a bug, because the mutation is unobservable.

(3) `-Wcxx23-compat-assign-to-class-rvalue`, which is off by default,
but *is* emitted in all C++ modes. Enabling this diagnostic will give
you a lot of noise from types like `tuple<int&>`, but you can opt into
it if you want.