[clang] -Wassign-to-class-rvalue
Implement two new warnings:
(1) `-Wassign-to-class-rvalue`, which is part of `-Wmost` and therefore
part of `-Wall`. This covers assignment, compound assignment (e.g. `+=`),
and postincrement and postdecrement.

(2) `-Wpreincrement-of-class-rvalue`, which is off by default.
This covers preincrement and predecrement. These do happen a lot in
real life; specifically with iterator types. For example ASIO contains
expressions of both forms: `--end()` and `++begin()`. These could be
rewritten to `std::prev(end())` and `std::next(begin())`, but in practice
I doubt people really want to do that.

On the other hand, `begin()++` or `end()--` is back in the category
of always being a bug, because the mutation is unobservable.

Unfortunately `-Wassign-to-class-rvalue` gives false positives on
simple types like `tuple<int&>` in C++20; it really "works" only in
C++23 and later.