[libc++] Add constexpr to std::relocate_at
One caveat: Suppose T is trivially relocatable but not nothrow-move-constructible.
Then we will memcpy it at runtime but move-construct it at constexpr time.
So `std::relocate_at` wants to say it's noexcept at runtime, but at constexpr time
it legitimately might throw an exception. C++26 or '29 will likely support
throwing exceptions at constexpr time, so, we can't fudge this. Leave
`std::relocate_at` as non-noexcept in this scenario for now.

Theoretically we could rely on Corentin's Clang builtin
`__builtin_trivially_relocate` to move-and-destroy the thing without throwing;
but I'm not sure that actually works in practice yet.