[libc++] Trivially relocatable (1/4): Traits and algorithms
This implements the library behavior proposed in P1144R9.
This patch doesn't do any special optimizations (e.g. to `std::vector`);
those will come later. This is just the bare minimum to conform to
P1144's additions to the standard.

    Feature-test macro `__cpp_lib_trivially_relocatable`

    concept relocatable

    struct is_relocatable<T> (*)
    struct is_nothrow_relocatable<T> (*)
    struct is_trivially_relocatable<T>

    bool is_relocatable_v<T> (*)
    bool is_nothrow_relocatable_v<T> (*)
    bool is_trivially_relocatable_v<T>

    T relocate(T*)
    T* relocate_at(T*, T*)
    ForwardIt uninitialized_relocate(InputIt, InputIt, ForwardIt)
    pair<InputIt, ForwardIt> uninitialized_relocate_n(InputIt, Size, ForwardIt)
    BidirIt2 uninitialized_relocate_backward(BidirIt1, BidirIt1, BidirIt2)

One caveat: Suppose T is trivially relocatable but not nothrow-move-constructible.
Then we will memcpy it at runtime but move-construct it at constexpr time.
So `std::relocate_at` wants to say it's noexcept at runtime, but at constexpr time
it legitimately might throw an exception. Leave `std::relocate_at` as non-noexcept
in this scenario for now.